Lower the effort level

+++ src/place/placers/analytical/GradientPlacer.java
@@ -92,7 +92,7 @@ public abstract class GradientPlacer extends AnalyticalAndGradientPlacer {
         options.add(
                 O_OUTER_EFFORT_LEVEL,
                 "number of solve-legalize iterations",
-                new Integer(30));
+                new Integer(27));
         options.add(
                 O_INNER_EFFORT_LEVEL_START,
                 "number of gradient steps to take in each outer iteration in the beginning",
